+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial stand up of repo and successful test deployment after extracting console-ui from main filmdrop terraform modules repository ([2](https://github.com/Element84/terraform-aws-filmdrop-ui/pull/2))
 - Renaming from `console-ui` and all references to more idiomatic `filmdrop-ui` to align with rest of filmdrop ecosystem
+- Fix an edge case of bucket name length being too long due to filmdrop inputs
 
 ### Changed
 
@@ -22,8 +23,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
-
-
 
 ### Changed